Module Cohttp_lwt.Client

The Make functor glues together a Cohttp.S.IO implementation to send requests down a connection that is established by the Net module. The resulting module satisfies the Client module type.

The Logs source name for this module's logger is "cohttp.lwt.client". When logging is enabled (at least warning level), eventual body leaks will be logged and easier to track.

module Make (Connection : S.Connection) : S.Client with type ctx = Connection.Net.ctx