Module Cohttp_lwt.Connection

exception Retry

Raised on failed requests that may be safely retried - even on non-idempotent requests. Raised for example on timeout or connection suhtdown by remote end.

module Make (Net : S.Net) : S.Connection with module Net = Net