Module Cohttp_lwt.Connection_cache

exception Retry

module Make_no_cache (Connection : S.Connection) : sig ... end

This functor establishes a new connection for each request.

module Make (Connection : S.Connection) (Sleep : S.Sleep) : sig ... end

This functor keeps a cache of connecions for reuse. Connections are reused based on their remote Conduit.endp (effectively IP / port).