Module Connection_cache.Make_no_cache

This functor establishes a new connection for each request.

Parameters

module Connection : S.Connection

Signature

include S.Connection_cache

type t

val call : t -> S.call

Process a request. Please see call.

val create : ?ctx:Connection.Net.ctx -> unit -> t

create ?ctx () creates a connection for handling a single request. The connection accepts only a single request and will automatically be closed as soon as possible.

parameter ctx
See

onnection.Net.ctx