Module type S.Server

The Server module implements a pipelined HTTP/1.1 server.

module IO : IO

type conn = IO.conn * Cohttp.Connection.t

type response_action = [

| `Expert of Cohttp.Response.t * (IO.ic -> IO.oc -> unit Lwt.t)

| `Response of Cohttp.Response.t * Body.t

]

A request handler can respond in two ways:

Using `Response, with a Response.t and a Body.t.

Using `Expert, with a Response.t and an IO function that is expected to write the response body. The IO function has access to the underlying IO.ic and IO.oc, which allows writing a response body more efficiently, stream a response or to switch protocols entirely (e.g. websockets). Processing of pipelined requests continue after the unit Lwt.t is resolved. The connection can be closed by closing the IO.ic.

type t

val make_response_action : ?⁠conn_closed:(conn -> unit) -> callback:(conn -> Cohttp.Request.t -> Body.t -> response_action Lwt.t) -> unit -> t

val make_expert : ?⁠conn_closed:(conn -> unit) -> callback:(conn -> Cohttp.Request.t -> Body.t -> (Cohttp.Response.t * (IO.ic -> IO.oc -> unit Lwt.t)) Lwt.t) -> unit -> t

val make : ?⁠conn_closed:(conn -> unit) -> callback:(conn -> Cohttp.Request.t -> Body.t -> (Cohttp.Response.t * Body.t) Lwt.t) -> unit -> t

val resolve_local_file : docroot:string -> uri:Uri.t -> string

Resolve a URI and a docroot into a concrete local filename.

val respond : ?⁠headers:Cohttp.Header.t -> ?⁠flush:bool -> status:Cohttp.Code.status_code -> body:Body.t -> unit -> (Cohttp.Response.t * Body.t) Lwt.t

respond ?headers ?flush ~status ~body will respond to an HTTP request with the given status code and response body. If flush is true, then every response chunk will be flushed to the network rather than being buffered. flush is true by default. The transfer encoding will be detected from the body value and set to chunked encoding if it cannot be determined immediately. You can override the encoding by supplying an appropriate Content-length or Transfer-encoding in the headers parameter.

val respond_string : ?⁠flush:bool -> ?⁠headers:Cohttp.Header.t -> status:Cohttp.Code.status_code -> body:string -> unit -> (Cohttp.Response.t * Body.t) Lwt.t

val respond_error : ?⁠headers:Cohttp.Header.t -> ?⁠status:Cohttp.Code.status_code -> body:string -> unit -> (Cohttp.Response.t * Body.t) Lwt.t

val respond_redirect : ?⁠headers:Cohttp.Header.t -> uri:Uri.t -> unit -> (Cohttp.Response.t * Body.t) Lwt.t

val respond_need_auth : ?⁠headers:Cohttp.Header.t -> auth:Cohttp.Auth.challenge -> unit -> (Cohttp.Response.t * Body.t) Lwt.t

val respond_not_found : ?⁠uri:Uri.t -> unit -> (Cohttp.Response.t * Body.t) Lwt.t

val callback : t -> IO.conn -> IO.ic -> IO.oc -> unit Lwt.t