Module Cohttp_lwt__Make.Response

Parameters

IO : Cohttp_lwt.S.IO

Signature

include Cohttp.Response

This contains the metadata for a HTTP/1.1 response header, including the encoding, headers, version, status code and whether to flush the connection after every body chunk (useful for server-side events and other long-lived connection protocols). The body is handled by the separate S module type, as it is dependent on the IO implementation.

The interface exposes a fieldslib interface which provides individual accessor functions for each of the records below. It also provides sexp serializers to convert to-and-from an Core.Std.Sexp.t.

include Cohttp.S.Response

type t = {

encoding : Cohttp.Transfer.encoding;
Transfer encoding of this HTTP response

headers : Cohttp.Header.t;
response HTTP headers

version : Cohttp.Code.version;
(** HTTP version, usually 1.1 *)

status : Cohttp.Code.status_code;
HTTP status code of the response

flush : bool;

}

include Ppx_sexp_conv_lib.Sexpable.S with type Response.t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val encoding : t -> Cohttp.Transfer.encoding

val headers : t -> Cohttp.Header.t

val version : t -> Cohttp.Code.version

val status : t -> Cohttp.Code.status_code

val flush : t -> bool

val compare : t -> t -> int

val make : ?⁠version:Cohttp.Code.version -> ?⁠status:Cohttp.Code.status_code -> ?⁠flush:bool -> ?⁠encoding:Cohttp.Transfer.encoding -> ?⁠headers:Cohttp.Header.t -> unit -> t

The response creates by make ~encoding ~headers () has an encoding value determined from the content of headers or if no proper header is present, using the value of encoding. Checked headers are "content-length", "content-range" and "transfer-encoding". The default value of encoding is chunked.

val pp_hum : Stdlib.Format.formatter -> t -> unit

Human-readable output, used by the toplevel printer

module Make = Cohttp.Response.Make

Functor to construct the IO-specific response handling function

include module type of sig ... end with type t := t

type t = Cohttp__Response.t

type reader = Cohttp__Response.Make(IO).reader

type writer = Cohttp__Response.Make(IO).writer

module IO : sig ... end

val read : IO.ic -> [ `Eof | `Invalid of string | `Ok of t ] IO.t

val has_body : t -> [ `No | `Unknown | `Yes ]

val make_body_writer : ?⁠flush:bool -> t -> IO.oc -> writer

val make_body_reader : t -> IO.ic -> reader

val read_body_chunk : reader -> Cohttp.Transfer.chunk IO.t

val write_header : t -> IO.oc -> unit IO.t

val write_body : writer -> string -> unit IO.t

val write : ?⁠flush:bool -> (writer -> unit IO.t) -> t -> IO.oc -> unit IO.t