Module Cohttp_lwt

module type IO = S.IO

module Request = Cohttp.Request

module Response = Cohttp.Response

module Make_client = Cohttp_lwt__.Client.Make

module Make_server = Cohttp_lwt__.Server.Make

module S : sig ... end

Portable Lwt implementation of HTTP client and server, without depending on a particular I/O implementation. The various Make functors must be instantiated by an implementation that provides a concrete IO monad.

module Body : sig ... end