Module Cohttp_lwt

module type IO = S.IO

module Request = Cohttp.Request

module Response = Cohttp.Response

module Connection : sig ... end

module Connection_cache : sig ... end

module Client : sig ... end

The Make functor glues together a Cohttp.S.IO implementation to send requests down a connection that is established by the Net module. The resulting module satisfies the Client module type.

module Server : sig ... end

The Make functor glues together a Cohttp.S.IO implementation to send requests down a connection that is established by the user. The resulting module satisfies the Server module type.

module Make_client (IO : IO) (Net : S.Net with module IO = IO) : sig ... end

module Make_server = Server.Make

module S : sig ... end

Portable Lwt implementation of HTTP client and server, without depending on a particular I/O implementation. The various Make functors must be instantiated by an implementation that provides a concrete IO monad.

module Body : sig ... end

module Private : sig ... end