Module Cohttp_lwt.Connection_cache

module Make_no_cache (Connection : S.Connection) : sig ... end

This functor establishes a new connection for each request.

module Make (Connection : S.Connection) (Sleep : S.Sleep) : sig ... end

This functor keeps a cache of connections for reuse. Connections are reused based on their remote Conduit.endp (effectively IP / port).

module Make_proxy (Connection : S.Connection) (Sleep : S.Sleep) : sig ... end

This functor keeps a cache of connections for reuse. Connections are reused based on their remote Conduit.endp (effectively IP / port). It also supports automatically connecting and reconnecting to direct and tunneling proxies, based on the remote URI scheme (HTTP will select direct proxies, HTTPS tunneling proxies).