Module type S.Net

The Net module type defines how to connect to a remote node and close the resulting channels to clean up.

module IO : IO

type ctx

val sexp_of_ctx : ctx -> Sexplib0.Sexp.t

val default_ctx : ctx

val connect_uri : ctx:ctx -> Uri.t -> (IO.conn * IO.ic * IO.oc) Lwt.t

val close_in : IO.ic -> unit

val close_out : IO.oc -> unit

val close : IO.ic -> IO.oc -> unit